+++ Ranger/UI/PlayersTableViewController/Constants.swift
@@ -1,0 +1,9 @@
+//
+//  Constants.swift
+//  Ranger
+//
+//  Created by Geri Borbás on 2020. 01. 18..
+//  Copyright © 2020. Geri Borbás. All rights reserved.
+//
+
+import Foundation